generate_svd: be more general with DTS fields
Generalizes the DTS parser to not rely on the presence of the
`reg-names` field to parse `reg` field in peripherals child nodes.

Instead, rely on the structure as defined by the DTS specification, i.e.
the address field is first indexed at `<address-cells> - 1`, and the
size field is second indexed at `<address-cells> + <size-cells> - 1`.

Search compatible list for `clic0`, `clint0`, and `plic0` strings, and
modifies the `script_path` to point to the relevant regmap generator.