fixup: scripts: add default `resetValue`
Adds a default `resetValue` field to registers defined with
`starfive_common/generate_register`.

Allows SVD parsers to supply a default reset value for the register.

Defaults to zero, which should be a safe reset value for most registers.
Users that require a non-zero default can supply it as an argument to
`generate_register`.